move a prop type to a file
