+++ src/utils/data.js
@@ -1,5 +1,3 @@
-import PropTypes from 'prop-types';
-
 const testData = [
     {
        "_id":"60666c42cc7b410027a1a9b1",
@@ -213,13 +211,4 @@ const testData = [
     }
  ];
 
-const ingredientPropTypes = PropTypes.shape({
-    _id: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-    type: PropTypes.oneOf(['bun', 'main', 'sauce']).isRequired,
-    price: PropTypes.number.isRequired,
-    image: PropTypes.string.isRequired,
-    image_mobile: PropTypes.string.isRequired,
-});
-
-export {testData, ingredientPropTypes};
+export {testData};
