+++ src/utils/prop-type.js
@@ -1,0 +1,12 @@
+import PropTypes from 'prop-types';
+
+const ingredientPropTypes = PropTypes.shape({
+    _id: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+    type: PropTypes.oneOf(['bun', 'main', 'sauce']).isRequired,
+    price: PropTypes.number.isRequired,
+    image: PropTypes.string.isRequired,
+    image_mobile: PropTypes.string.isRequired,
+});
+
+export {ingredientPropTypes};
